requirements.txt
flask==3.1.2
sqlalchemy==2.0.43
python-dotenv==1.1.1
sqlalchemy-libsql==0.2.0
pymysql==1.1.1
